requirements.txt
django
pillow
djangorestframework
drf_spectacular
django-filters
